refactor(gateway): keep catalog guards within budgets
